qwen_service/requirements.txt
accelerate==1.1.1
fastapi==0.115.0
pillow==10.4.0
pydantic==2.9.2
sentencepiece==0.2.0
torch==2.5.1
torchvision==0.20.1
transformers==4.49.0
uvicorn[standard]==0.30.6
